+++ src/rest/v2/clients/base.ts
@@ -56,6 +56,8 @@ export interface PatreonClientOptions<IncludeAll extends boolean = false> {
     /**
      * Options for storing and getting API (creator) tokens.
      * @default undefined
+     * @deprecated Use the `CacheTokenStore` class to store and fetch any tokens.
+     * See the documentation for an example: https://patreon-api.pages.dev/guide/features/oauth#creator-token
      */
     store?: PatreonTokenFetchOptions
 }
@@ -70,8 +72,20 @@ export abstract class PatreonClient<IncludeAll extends boolean = false> extends 
      */
     public webhooks: WebhookClient
 
+    /**
+     * Applies the `simplify` method on all responses.
+     *
+     * NOTE: due to type complexity, the response might be typed as an union of a paginated and resource response.
+     * If this happens, apply the `simplify` method on the default response manually. I have no idea how to fix this bug.
+     */
     public simplified: PatreonSharedClient<'simplified', IncludeAll>
 
+    /**
+     * Applies the `normalize` method on all responses.
+     *
+     * NOTE: due to type complexity, the response might be typed as an union of a paginated and resource response.
+     * If this happens, apply the `normalize` method on the default response manually. I have no idea how to fix this bug.
+     */
     public normalized: PatreonSharedClient<'normalized', IncludeAll>
 
     // eslint-disable-next-line jsdoc/require-returns
@@ -155,6 +169,8 @@ export abstract class PatreonClient<IncludeAll extends boolean = false> extends 
     /**
      * Fetch the stored token with the `get` method from the client options
      * @returns the stored token, if `options.store.get` is defined and returns succesfully.
+     * @deprecated Use the `CacheTokenStore` class to store and fetch any tokens.
+     * See the documentation for an example: https://patreon-api.pages.dev/guide/features/oauth#creator-token
      */
     public async fetchStoredToken(): Promise<Oauth2StoredToken | undefined> {
         return PatreonClient.fetchStored(this.store)
@@ -164,6 +180,8 @@ export abstract class PatreonClient<IncludeAll extends boolean = false> extends 
      * Save your token with the method from the client options
      * @param token The token to save
      * @param [cache] Whether to overwrite the application token cache and update it with the token
+     * @deprecated Use the `CacheTokenStore` class to store and fetch any tokens.
+     * See the documentation for an example: https://patreon-api.pages.dev/guide/features/oauth#creator-token
      */
     public async putStoredToken(token: Oauth2StoredToken, cache?: boolean): Promise<void> {
         await this.store?.put(token)